split buildtime and runtime in config file
