+++ config.test.json
@@ -1,28 +1,29 @@
 {
-  "api": {
-    "protocol": "https",
-    "host": "www.api.dev.meetup.com",
-    "timeout": 10,
-    "root_url": "https://www.api.dev.meetup.com"
+  "buildtime": {
+    "asset_server": {
+      "host": "beta2.dev.meetup.com",
+      "port": 0
+    },
+    "app_server": {
+      "host": "beta2.dev.meetup.com",
+      "port": 0
+    }
   },
-  "asset_server": {
-    "host": "beta2.dev.meetup.com",
-    "port": 0
-  },
-  "csrf_secret": "asdfasdfasdfasdfasdfasdfasdfasdf",
-  "cookie_encrypt_secret": "asdfasdfasdfasdfasdfasdfasdfasdf",
-  "app_server": {
-    "host": "beta2.dev.meetup.com",
-    "port": 0
-  },
-  "duotone_urls": [
-    "http://example.com/duotone.jpg"
-  ],
-  "photo_scaler_salt": "asdfasdfasdfasdfasdfasdfasdfasdf",
-  "oauth": {
-    "auth_url": "https://secure.dev.meetup.com/oauth2/authorize",
-    "access_url": "https://secure.dev.meetup.com/oauth2/access",
-    "key": "asdfasdfasdfasdfasdfasdfasdfasdf",
-    "secret": "asdfasdfasdfasdfasdfasdfasdfasdf"
+  "runtime": {
+    "api": {
+      "protocol": "https",
+      "host": "www.api.dev.meetup.com",
+      "timeout": 10,
+      "root_url": "https://www.api.dev.meetup.com"
+    },
+    "csrf_secret": "asdfasdfasdfasdfasdfasdfasdfasdf",
+    "cookie_encrypt_secret": "asdfasdfasdfasdfasdfasdfasdfasdf",
+    "photo_scaler_salt": "asdfasdfasdfasdfasdfasdfasdfasdf",
+    "oauth": {
+      "auth_url": "https://secure.dev.meetup.com/oauth2/authorize",
+      "access_url": "https://secure.dev.meetup.com/oauth2/access",
+      "key": "asdfasdfasdfasdfasdfasdfasdfasdf",
+      "secret": "asdfasdfasdfasdfasdfasdfasdfasdf"
+    }
   }
 }
