+++ src/util/config/index.js
@@ -2,7 +2,8 @@ import fs from 'fs';
 import convict from 'convict';
 import path from 'path';
 
-import { duotones, getDuotoneUrls } from './duotone';
+import { schema as buildSchema, validateProtocol } from './build';
+import { duotones, getDuotoneUrls } from '../duotone';
 
 /**
  * This module provides a single source of truth for application configuration
@@ -31,19 +32,12 @@ import { duotones, getDuotoneUrls } from './duotone';
 
 const random32 = 'asdfasdfasdfasdfasdfasdfasdfasdf';
 
-export const PROTOCOL_ERROR = 'Protocol must be http or https';
 export const COOKIE_SECRET_ERROR =
 	'Cookie Encrypt Secret must be a random 32+ char string';
 export const CSRF_SECRET_ERROR = 'CSRF Secret must be a random 32+ char string';
 export const OAUTH_SECRET_ERROR = 'Invalid OAUTH Secret';
 export const OAUTH_KEY_ERROR = 'Invalid OAUTH Key';
 export const SALT_ERROR = 'Invalid Photo Scaler Salt';
-
-export const validateProtocol = protocol => {
-	if (!['http', 'https'].includes(protocol)) {
-		throw new Error(PROTOCOL_ERROR);
-	}
-};
 
 export const validateCookieSecret = secret => {
 	if (!secret || secret.toString().length < 32) {
@@ -75,12 +69,8 @@ export const validatePhotoScalerSalt = salt => {
 	}
 };
 
-let config = convict({
-	env: {
-		format: ['production', 'development', 'test'],
-		default: 'development',
-		env: 'NODE_ENV',
-	},
+export const config = convict({
+	...buildSchema,
 	api: {
 		protocol: {
 			format: validateProtocol,
@@ -102,68 +92,21 @@ let config = convict({
 			default: '',
 		},
 	},
-	asset_server: {
-		host: {
-			format: String,
-			default: 'beta2.dev.meetup.com',
-			env: 'ASSET_SERVER_HOST',
-		},
-		path: {
-			format: String,
-			default: '/static',
-			env: 'ASSET_PATH',
-		},
-		port: {
-			format: 'port',
-			default: 8001,
-			arg: 'asset-port',
-			env: 'ASSET_SERVER_PORT',
-		},
-	},
-	disable_hmr: {
-		format: Boolean,
-		default: false,
-		env: 'DISABLE_HMR',
-	},
-	cookie_encrypt_secret: {
-		format: validateCookieSecret,
-		default: process.env.NODE_ENV !== 'production' && random32,
-		env: 'COOKIE_ENCRYPT_SECRET',
-	},
-	csrf_secret: {
-		format: validateCsrfSecret,
-		default: process.env.NODE_ENV !== 'production' && random32,
-		env: 'CSRF_SECRET',
-	},
-	app_server: {
-		protocol: {
-			format: validateProtocol,
-			default: 'http',
-			env: 'DEV_SERVER_PROTOCOL', // legacy naming
-		},
-		host: {
-			format: String,
-			default: 'beta2.dev.meetup.com',
-			env: 'DEV_SERVER_HOST', // legacy naming
-		},
-		port: {
-			format: 'port',
-			default: 8000,
-			arg: 'app-port',
-			env: 'DEV_SERVER_PORT', // legacy naming
-		},
-	},
 	duotone_urls: {
 		format: Object,
 		default: {},
 	},
-	isDev: {
-		format: Boolean,
-		default: true,
+	cookie_encrypt_secret: {
+		format: validateCookieSecret,
+		default: (process.env.NODE_ENV !== 'production' && random32) || '',
+		env: 'COOKIE_ENCRYPT_SECRET',
+		sensitive: true,
 	},
-	isProd: {
-		format: Boolean,
-		default: false,
+	csrf_secret: {
+		format: validateCsrfSecret,
+		default: (process.env.NODE_ENV !== 'production' && random32) || '',
+		env: 'CSRF_SECRET',
+		sensitive: true,
 	},
 	oauth: {
 		auth_url: {
@@ -180,38 +123,37 @@ let config = convict({
 			format: validateOauthSecret,
 			default: null,
 			env: 'MUPWEB_OAUTH_SECRET',
+			sensitive: true,
 		},
 		key: {
 			format: validateOauthKey,
 			default: null,
 			env: 'MUPWEB_OAUTH_KEY',
+			sensitive: true,
 		},
 	},
 	photo_scaler_salt: {
 		format: validatePhotoScalerSalt,
 		default: null,
 		env: 'PHOTO_SCALER_SALT',
+		sensitive: true,
 	},
 });
 
 // Load environment dependent configuration
-const configFile = path.resolve(
+const configPath = path.resolve(
 	process.cwd(),
 	`config.${config.get('env')}.json`
 );
 
-if (fs.existsSync(configFile)) {
-	config.loadFile(configFile);
+if (fs.existsSync(configPath)) {
+	const runConfig = require(configPath).runtime || {};
+	config.load(runConfig);
 }
 
 config.set(
 	'duotone_urls',
 	getDuotoneUrls(duotones, config.get('photo_scaler_salt'))
-);
-
-config.set(
-	'api.root_url',
-	`${config.get('api.protocol')}://${config.get('api.host')}`
 );
 
 config.set('isProd', config.get('env') === 'production');
